+++ pdfalyzer/util/adobe_strings.py
@@ -43,7 +43,7 @@ NAMED           = '/Named'
 NAMES           = '/Names'
 NEXT            = '/Next'
 NUMS            = '/Nums'
-OBJECT_STREAM   = '/ObjStm'
+OBJ_STM         = '/ObjStm'
 OBJ             = '/Obj'
 # TODO: /Pg refs could be the parents of /OBJR?
 OBJR            = '/OBJR'  # Object reference to "an entire PDF object"