+++ tests/conftest.py
@@ -38,6 +38,11 @@ def attachment_pdf_pdfalyzer():
 def geobase_pdfalyzer():
     return Pdfalyzer(str(FIXTURES_DIR.joinpath('GeoBase_NHNC1_Data_Model_UML_EN.pdf')))
 
+# Has a lot of unplaced nodes
+@pytest.fixture(scope='session')
+def form_evince_pdfalyzer():
+    return Pdfalyzer(str(FIXTURES_DIR.joinpath('form_evince.pdf')))
+
 
 # Some obj ids for use with -f when you want to limit yourself to the font
 @pytest.fixture(scope="session")